Center menu

+++ qt-qml-gui/piecesmodel.cpp
@@ -53,7 +53,8 @@ void PiecesModel::newGame(gsl::not_null<ChessGame*> game)
     auto board = gameAccess->get_board();
 
     if (myPieces.count() > 0) {
-        beginRemoveRows(QModelIndex{}, 0, myPieces.count() - 1);
+        beginRemoveRows(QModelIndex{},  0,
+                        gsl::narrow<int>(myPieces.count() - 1));
         myPieces.clear();
         endRemoveRows();
     }
